Move IpAddress to uri/tools so it can be reused
Rather than having multiple separate implementations for parsing and
converting IP addresses between representations, we should reuse the
existing IpAddress class anywhere that conversions are done (e.g.
BuildUAuthority). Moving the IpAddress.h file to uri/tools makes it
clear this is an independent tool and not *specifically* part of the
serializer, and makes the dependency graph between uri submodules clear.
