+++ include/up-cpp/uri/serializer/MicroUriSerializer.h
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2023 General Motors GTO LLC
+ * Copyright (c) 2024 General Motors GTO LLC
  *
  * Licensed to the Apache Software Foundation (ASF) under one
  * or more contributor license agreements.  See the NOTICE file
@@ -19,7 +19,7 @@
  * under the License.
  * 
  * SPDX-FileType: SOURCE
- * SPDX-FileCopyrightText: 2023 General Motors GTO LLC
+ * SPDX-FileCopyrightText: 2024 General Motors GTO LLC
  * SPDX-License-Identifier: Apache-2.0
  */
 #ifndef MICRO_URI_SERIALIZER_H_
@@ -29,7 +29,7 @@
 #include <up-cpp/uri/builder/BuildUAuthority.h>
 #include <up-cpp/uri/builder/BuildEntity.h>
 #include <up-cpp/uri/builder/BuildUResource.h>
-#include "up-cpp/uri/serializer/IpAddress.h"
+#include "up-cpp/uri/tools/IpAddress.h"
 
 using AddressType = uprotocol::uri::IpAddress::AddressType;
