+++ src/uri/MicroUriSerializer.cpp
@@ -25,7 +25,7 @@
 
 
 #include <up-cpp/uri/serializer/MicroUriSerializer.h>
-#include <up-cpp/uri/serializer/IpAddress.h>
+#include <up-cpp/uri/tools/IpAddress.h>
 
 using uprotocol::uri::IpAddress;
 using AddressType = IpAddress::AddressType;
