+++ test/uri/serializer/MicroUriSerializerTest.cpp
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2023 General Motors GTO LLC
+ * Copyright (c) 2024 General Motors GTO LLC
  *
  * Licensed to the Apache Software Foundation (ASF) under one
  * or more contributor license agreements.  See the NOTICE file
@@ -19,14 +19,14 @@
  * under the License.
  * 
  * SPDX-FileType: SOURCE
- * SPDX-FileCopyrightText: 2023 General Motors GTO LLC
+ * SPDX-FileCopyrightText: 2024 General Motors GTO LLC
  * SPDX-License-Identifier: Apache-2.0
  */
 #include <string>
 #include <gtest/gtest.h>
 #include <up-cpp/uri/serializer/MicroUriSerializer.h>
 #include <up-cpp/uri/builder/BuildUUri.h>
-#include <up-cpp/uri/serializer/IpAddress.h>
+#include <up-cpp/uri/tools/IpAddress.h>
 #include <up-cpp/uri/tools/Utils.h>
 
 using namespace uprotocol::uri;
